add notification system

+++ src/controllers/job.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from "express";
 import prisma from "../prisma";
+import { IoEvent } from "../routes/IoEventTypes";
 
 
 export const find: RequestHandler = async (req, res, next) => {
@@ -31,9 +32,25 @@ export const all: RequestHandler = async (req, res, next) => {
 
 export const destroy: RequestHandler = async (req, res, next) => {
     try {
+        const job = await prisma.job.findUnique({
+            where: { id: req.params.id },
+        });
+        if (!job) {
+            return res.status(204).send();
+        }
+        if (job.userId !== req.user!.id) {
+            return res.status(403).send();
+        }
         await prisma.job.delete({
             where: { id: req.params.id },
         });
+
+        res.notifications = [
+          {
+            message: { record: 'JOB', id: req.params.id },
+            event: IoEvent.DELETED_RECORD
+          }
+        ]
         res.status(204).send();
     } catch (error) {
         next(error);
